murmur: factor out murmur_round and murmur_finish
The goal is to allow incremental (aka streaming) hashing. With this API, the
caller doesn't need to have all the data in one big continuous buffer. Instead,
it can be hashed as it's read from disk or a socket.

The usecase I have in mind is the IVS tcam module, which needs to mask and hash
packet keys. It's faster to fuse those loops together instead of producing a
temporary masked buffer and then hashing that.