slot_allocator: add hole limiting
A hole is defined as a slot that is both free and smaller than alloc_start.  We
normally search for free slots starting from the last slot we allocated.  This
is fast, but has the downside of spreading allocations over all of the backing
memory even when the number of allocated objects at any time is much smaller.

With this change, once the number of holes becomes greater than a constant
fraction (1%) of the total number of slots, we reset the allocation clock.
Because we know >1% of the slots before alloc_start are free, the allocation
time is still amortized constant time. We'd expect to only have to look through
4 bitmap words (128 bits) to find the next free slot.