True Relative Semaphore Timeout Support.
The posix semaphore specification (inexplicably) requires timeout values relative to the REALTIME clock.
Changes to the value of the REALTIME while waiting on the semaphore (due to manual or NTP updates)
will make the timeout expire prematurely if the change is in the future or wildly overdue if the change is in the past.

The os_sem_create() function has been augmented with a backwards-compatible, optional 'flags' argument.
The only flag current supported is OS_SEM_CREATE_F_TRUE_RELATIVE_TIMEOUTS.

When the OS_SEM_CREATE_F_TRUE_RELATIVE_TIMEOUTS flags is specified at creation time, semaphores
will be implemented internally using the eventfd interface (instead of the posix interface).
Timeouts are then implemented with poll(), whose timeout value is relative and immune to REALTIME clock changes.

All clients which rely on os_sem_take_timeout() should use the OS_SEM_CREATE_F_TRUE_RELATIVE_TIMEOUTS flag.
Clients which do not use os_sem_take_timeout() on the semaphore need not specify it (and probably shouldn't due to performance differences).